Added fix in loading script for cases where match ID already exists in the database. In such cases, the code will no longer error out. The corresponding file will be deleted.

+++ pydantic_models/valorant_match_api_pydantic_model.py
@@ -114,8 +114,8 @@ class BehaviorFactors(BaseModel):
     collisions: float
     commsRatingRecovery: int
     damageParticipationOutgoing: int
-    friendlyFireIncoming: int
-    friendlyFireOutgoing: int
+    friendlyFireIncoming: float
+    friendlyFireOutgoing: float
     mouseMovement: int
     selfDamage: float
     stayedInSpawnRounds: float
